Add files via upload
This is the current set up for my Tevo Tarantula Original with a 32 bit e3 mini 2.0 and tft 3.5 v3.0.
I've updated my screen firmware to latest. 

There are a few tweaks that I am planning - the z doesn't lift after homing, the jerk and acceleration are to be tuned.

The hot end is for an e3d v6. Change the hotend value to match yours (1 for stock (I believe) in line 422). 

The extruder is set for a bmg geared extruder. Change this to match your filament extrusion (e0 was 100 with the stock extruder, as commented, line 762). 

PID control is used for both bed heating and hot end. Your hot end pid will need to be tuned (lines 495 - 297). The bed PID seems to work well with the 12v standard bed and a 5mm piece of glass on top. 

Comment out the skew in the Configuration.h (lines 1449, 1461, 1462, 1473) otherwise you will have skew based on my frame that is still in the works of getting aligned. 

Your mileage may vary. This works well on my machine, but I assume no risk for its use.